refactor: moved Repositories intefaces to the corresponding folder [#88]
