+++ StudyConnect.API.Tests/Mock/CategoryControllerTests.cs
@@ -47,7 +47,7 @@ public class CategoryControllerTests
         };
 
         _mockRepo.Setup(repo => repo.GetCategoryByIdAsync(It.IsAny<Guid>()))
-                 .ReturnsAsync(OperationResult<ForumCategory?>.Success(testCategory));
+                 .ReturnsAsync(OperationResult<ForumCategory>.Success(testCategory));
 
         // Act
         var result = await _controller.GetCategoryById(Guid.NewGuid());
@@ -63,7 +63,7 @@ public class CategoryControllerTests
     {
         // Arrange
         _mockRepo.Setup(repo => repo.GetCategoryByIdAsync(It.IsAny<Guid>()))
-                 .ReturnsAsync(OperationResult<ForumCategory?>.Failure(CategoryNotFound));
+                 .ReturnsAsync(OperationResult<ForumCategory>.Failure(CategoryNotFound));
 
         // Act
         var result = await _controller.GetCategoryById(Guid.NewGuid());
@@ -84,7 +84,7 @@ public class CategoryControllerTests
             };
 
         _mockRepo.Setup(repo => repo.GetAllCategoriesAsync())
-             .ReturnsAsync(OperationResult<IEnumerable<ForumCategory>>.Success(categories));
+             .ReturnsAsync(OperationResult<IEnumerable<ForumCategory?>>.Success(categories));
 
         // Act
         var result = await _controller.GetAllCategories();
@@ -113,7 +113,7 @@ public class CategoryControllerTests
     {
         // Arrange
         _mockRepo.Setup(repo => repo.GetAllCategoriesAsync())
-                 .ReturnsAsync(OperationResult<IEnumerable<ForumCategory>>.Failure("Something went wrong"));
+                 .ReturnsAsync(OperationResult<IEnumerable<ForumCategory?>>.Failure("Something went wrong"));
 
         // Act
         var result = await _controller.GetAllCategories();
