+++ StudyConnect.Data/Repositories/GroupRepository.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.EntityFrameworkCore;
 using StudyConnect.Core.Common;
-using StudyConnect.Core.Interfaces;
+using StudyConnect.Core.Interfaces.Repositories;
 using StudyConnect.Core.Models;
 
 namespace StudyConnect.Data.Repositories;
